LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/view-timeline-source.tentative-expected.txt


FAIL Default source for a View timeline is the nearest scroll ancestor to the subject assert_equals: expected "outer" but got "inner"
PASS View timeline ignores explicitly set source
FAIL View timeline source is null when display:none assert_equals: expected null but got Element node <div id="inner" class="scroller">
      <div id="target" ...
